# docs/openapi.yaml
openapi: 3.1.0
info:
  title: REST API Системы управления банковскими картами
  version: v1
servers:
- url: http://localhost:8080
  description: Generated server url
security:
- bearerAuth: []
...
paths:
  /api/v1/bank-cards/{cardId}/request-block/:
    put:
      tags:
      - Пользовательские методы для карт
      summary: Запрос на блокировку карты
      operationId: requestBlockMyCard
      parameters:
      - name: cardId
        in: path
        required: true
        schema:
          type: integer
          format: int64
      responses:
        "200":
          description: OK
          content:
            '*/*':
              schema:
                $ref: "#/components/schemas/BankCardDto"
  /api/v1/bank-cards/transfer:
    put:
      tags:
      - Пользовательские методы для карт
      summary: Перевод с одной на другую карту
      operationId: transfer
      requestBody:
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/TransferRequestDto"
        required: true
      responses:
        "200":
          description: OK
          content:
            '*/*':
              schema:
                $ref: "#/components/schemas/TransferResponseDto"
...
  /api/v1/bank-cards:
    get:
      tags:
      - Пользовательские методы для карт
      summary: Список моих карт
      operationId: getMyCards
      parameters:
      - name: pageable
        in: query
        required: true
        schema:
          $ref: "#/components/schemas/Pageable"
      - name: cardNumberLast4
        in: query
        required: false
        schema:
          type: string
          default: ""
      responses:
        "200":
          description: OK
          content:
            '*/*':
              schema:
                $ref: "#/components/schemas/PageBankCardDto"
  /api/v1/bank-cards/{cardId}/details:
    get:
      tags:
      - Пользовательские методы для карт
      summary: Получить скрытые данные моей карты
      operationId: getFullDetails
      parameters:
      - name: cardId
        in: path
        required: true
        schema:
          type: integer
          format: int64
      responses:
        "200":
          description: OK
          content:
            '*/*':
              schema:
                $ref: "#/components/schemas/BankCardDetailsDto"
...
components:
  schemas:
    BankCardDto:
      type: object
      properties:
        id:
          type: integer
          format: int64
        cardNumber:
          type: string
        expirationDate:
          type: string
          example: 2025-03
        status:
          type: string
          enum:
          - ACTIVE
          - BLOCKED
          - EXPIRED
        balance:
          type: number
        cardHolderName:
          type: string
    TransferRequestDto:
      type: object
      properties:
        fromCardNumber:
          type: string
          minLength: 1
        toCardNumber:
          type: string
          minLength: 1
        amount:
          type: number
          minimum: 0.01
      required:
      - amount
    TransferResponseDto:
      type: object
      properties:
        fromCard:
          $ref: "#/components/schemas/BankCardDto"
        toCard:
          $ref: "#/components/schemas/BankCardDto"
...
    Pageable:
      type: object
      properties:
        page:
          type: integer
          format: int32
          minimum: 0
        size:
          type: integer
          format: int32
          minimum: 1
        sort:
          type: array
          items:
            type: string
    PageBankCardDto:
      type: object
      properties:
        totalElements:
          type: integer
          format: int64
        totalPages:
          type: integer
          format: int32
        pageable:
          $ref: "#/components/schemas/PageableObject"
        size:
          type: integer
          format: int32
        content:
          type: array
          items:
            $ref: "#/components/schemas/BankCardDto"
        number:
          type: integer
          format: int32
        sort:
          $ref: "#/components/schemas/SortObject"
        first:
          type: boolean
        last:
          type: boolean
        numberOfElements:
          type: integer
          format: int32
        empty:
          type: boolean
    PageableObject:
      type: object
      properties:
        paged:
          type: boolean
        pageNumber:
          type: integer
          format: int32
        pageSize:
          type: integer
          format: int32
        unpaged:
          type: boolean
        offset:
          type: integer
          format: int64
        sort:
          $ref: "#/components/schemas/SortObject"
    SortObject:
      type: object
      properties:
        sorted:
          type: boolean
        unsorted:
          type: boolean
        empty:
          type: boolean
    BankCardDetailsDto:
      type: object
      properties:
        id:
          type: integer
          format: int64
        cardNumber:
          type: string
        expirationDate:
          type: string
          example: 2025-03
        status:
          type: string
          enum:
          - ACTIVE
          - BLOCKED
          - EXPIRED
        balance:
          type: number
        cardHolderName:
          type: string
...
  securitySchemes:
    bearerAuth:
      type: http
      in: header
      scheme: bearer
      bearerFormat: JWT